Support for parallel message generation and basic statistics
The generator now takes "tasks" parameter which tells it how many tasks it should launch. The server counts the number of send_message calls it has processed and the generator periodically retrieves this value and reports how many such calls have been processed in one millisecond.

The goal of this is to enable performance assessment of the proxide proxy.

Includes linting support in the pipeline.