Implemented basic verification unit test for proxide
The first unit test for proxide is now implemented with the help of the grpc_tester library which enables the unit test to start and stop gRPC server and gRPC message generation client.

The new server and generator allows the test to check whether proxide is able to intercept messages. In this first iteration the test only verifies that this basic principle works.