Converted unwrap to expect
Proxide panics when launched from RustRover. I tried to locate the source of the panic. Though it always seemed to come from the internals of the Rust main runtime. Launching the application in debug mode does not trigger the panic.

The application does not panic if a pause of 500 milliseconds is added before the "let chunk = f.size();" gets executed in marcos.rs:draw_views function.  The pause can be anywhere between the first line in main() and the function mentioned above.

unwrap -> expect conversions were done while hunting for this issue.